+++ src/main/java/com/saintdan/framework/component/LogHelper.java
@@ -1,9 +1,9 @@
 package com.saintdan.framework.component;
 
+import com.saintdan.framework.domain.LogDomain;
 import com.saintdan.framework.enums.LogType;
 import com.saintdan.framework.param.LogParam;
 import com.saintdan.framework.po.User;
-import com.saintdan.framework.domain.LogDomain;
 import com.saintdan.framework.tools.SpringSecurityUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;