+++ src/main/java/com/saintdan/framework/config/package-info.java
@@ -1,5 +1,16 @@
 /**
- * Java config files.
+ * Package of Java config.
+ *
+ * <pre>
+ *     {@link com.saintdan.framework.config.OAuth2ServerConfiguration}
+ *     Oauth2 server config.
+ *
+ *     {@link com.saintdan.framework.config.WebInitializer}
+ *     Spring boot servlet initializer.
+ *
+ *     {@link com.saintdan.framework.config.WebSecurityConfiguration}
+ *     Spring web security config.
+ * </pre>
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 6/23/15