+++ src/main/java/com/saintdan/framework/controller/ResourceController.java
@@ -7,16 +7,16 @@ import com.saintdan.framework.constant.CommonsConstant;
 import com.saintdan.framework.constant.ControllerConstant;
 import com.saintdan.framework.constant.ResourceURL;
 import com.saintdan.framework.constant.ResultConstant;
+import com.saintdan.framework.domain.ResourceDomain;
 import com.saintdan.framework.enums.ErrorType;
 import com.saintdan.framework.enums.OperationStatus;
 import com.saintdan.framework.exception.CommonsException;
 import com.saintdan.framework.param.ResourceParam;
 import com.saintdan.framework.po.User;
-import com.saintdan.framework.domain.ResourceDomain;
 import com.saintdan.framework.vo.ResultVO;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.validation.BindingResult;
@@ -28,7 +28,7 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 
 /**
- * Resource's controller.
+ * Controller of resource.
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 10/17/15
@@ -52,7 +52,7 @@ public class ResourceController {
     public ResultVO create(@CurrentUser User currentUser, @Valid ResourceParam param, BindingResult result, @PathVariable String sign) {
         try {
             // Validate current user, param and sign.
-            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, log);
+            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -60,10 +60,10 @@ public class ResourceController {
             return resourceService.create(param, currentUser);
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -77,17 +77,17 @@ public class ResourceController {
         try {
             ResourceParam param = new ResourceParam();
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
             return resourceService.getAllResources();
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -106,17 +106,17 @@ public class ResourceController {
             }
             ResourceParam param = new ResourceParam();
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
             return resourceService.getPage(new PageRequest(Integer.valueOf(pageNo), CommonsConstant.PAGE_SIZE));
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -134,17 +134,17 @@ public class ResourceController {
             }
             ResourceParam param = new ResourceParam(Long.valueOf(id));
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
             return resourceService.getResourceById(param);
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -162,7 +162,7 @@ public class ResourceController {
                 return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
             }
             // Validate current user, param and sign.
-            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, log);
+            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -170,10 +170,10 @@ public class ResourceController {
             return resourceService.update(param, currentUser);
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -191,7 +191,7 @@ public class ResourceController {
             }
             ResourceParam param = new ResourceParam(Long.valueOf(id));
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -201,10 +201,10 @@ public class ResourceController {
             return new ResultVO(ResultConstant.OK, OperationStatus.SUCCESS, String.format(ControllerConstant.DELETE, ROLE));
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -212,7 +212,7 @@ public class ResourceController {
     // PRIVATE FIELDS
     // ------------------------
 
-    private static final Log log = LogFactory.getLog(ResourceController.class);
+    private static final Logger logger = LoggerFactory.getLogger(ClientController.class);
 
     @Autowired
     private ResultHelper resultHelper;