+++ src/main/java/com/saintdan/framework/controller/RoleController.java
@@ -7,17 +7,17 @@ import com.saintdan.framework.constant.CommonsConstant;
 import com.saintdan.framework.constant.ControllerConstant;
 import com.saintdan.framework.constant.ResourceURL;
 import com.saintdan.framework.constant.ResultConstant;
+import com.saintdan.framework.domain.RoleDomain;
 import com.saintdan.framework.enums.ErrorType;
 import com.saintdan.framework.enums.OperationStatus;
 import com.saintdan.framework.exception.CommonsException;
 import com.saintdan.framework.param.RoleParam;
 import com.saintdan.framework.po.User;
-import com.saintdan.framework.domain.RoleDomain;
 import com.saintdan.framework.vo.ResultVO;
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.validation.BindingResult;
@@ -29,7 +29,7 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 
 /**
- * Role's controller.
+ * Controller of role.
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 10/17/15
@@ -53,7 +53,7 @@ public class RoleController {
     public ResultVO create(@CurrentUser User currentUser, @Valid RoleParam param, BindingResult result, @PathVariable String sign) {
         try {
             // Validate current user, param and sign.
-            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, log);
+            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -61,10 +61,10 @@ public class RoleController {
             return roleService.create(param, currentUser);
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -78,17 +78,17 @@ public class RoleController {
         try {
             RoleParam param = new RoleParam();
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
             return roleService.getAllRoles();
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -107,17 +107,17 @@ public class RoleController {
             }
             RoleParam param = new RoleParam();
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
             return roleService.getPage(new PageRequest(Integer.valueOf(pageNo), CommonsConstant.PAGE_SIZE));
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -135,17 +135,17 @@ public class RoleController {
             }
             RoleParam param = new RoleParam(Long.valueOf(id));
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
             return roleService.getRoleById(param);
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -163,7 +163,7 @@ public class RoleController {
                 return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
             }
             // Validate current user, param and sign.
-            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, log);
+            ResultVO resultVO = validateHelper.validate(result, currentUser, param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -171,10 +171,10 @@ public class RoleController {
             return roleService.update(param, currentUser);
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -194,7 +194,7 @@ public class RoleController {
             // Prepare to validate signature.
             param.setSign(new String(Base64.decodeBase64(sign.getBytes())));
             // Sign validate.
-            ResultVO resultVO = validateHelper.validate(param, sign, log);
+            ResultVO resultVO = validateHelper.validate(param, sign, logger);
             if (resultVO != null) {
                 return resultVO;
             }
@@ -204,10 +204,10 @@ public class RoleController {
             return new ResultVO(ResultConstant.OK, OperationStatus.SUCCESS, String.format(ControllerConstant.DELETE, ROLE));
         } catch (CommonsException e) {
             // Return error information and log the exception.
-            return resultHelper.infoResp(log, e.getErrorType());
+            return resultHelper.infoResp(logger, e.getErrorType());
         } catch (Exception e) {
             // Return unknown error and log the exception.
-            return resultHelper.errorResp(log, e, ErrorType.UNKNOWN, e.getMessage());
+            return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage());
         }
     }
 
@@ -215,7 +215,7 @@ public class RoleController {
     // PRIVATE FIELDS
     // ------------------------
 
-    private static final Log log = LogFactory.getLog(RoleController.class);
+    private static final Logger logger = LoggerFactory.getLogger(RoleController.class);
 
     @Autowired
     private ResultHelper resultHelper;